#116 [refactor-goods] goods Entity 리펙토링 및 GoodsInfo Entity 추가
