+++ src/main/java/com/sparta/ticketauction/domain/admin/request/PlaceRequest.java
@@ -1,12 +1,7 @@
 package com.sparta.ticketauction.domain.admin.request;
 
-import java.util.List;
-
 import com.sparta.ticketauction.domain.place.entity.Place;
-import com.sparta.ticketauction.domain.seat.request.SeatRequest;
 
-import jakarta.validation.Valid;
-import jakarta.validation.constraints.NotEmpty;
 import jakarta.validation.constraints.Size;
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
@@ -19,10 +14,6 @@ public class PlaceRequest {
 
 	@Size(min = 1, max = 150, message = "주소 입력은 필수입니다.")
 	private final String address;
-
-	@Valid
-	@NotEmpty(message = "좌석 정보는 필수입니다.")
-	private final List<SeatRequest> seats;
 
 	public Place toEntity(Integer countSeats) {
 		return Place
