+++ src/main/java/com/sparta/ticketauction/domain/goods/entity/Goods.java
@@ -1,16 +1,12 @@
 package com.sparta.ticketauction.domain.goods.entity;
 
 import java.time.LocalDate;
-import java.util.ArrayList;
-import java.util.List;
 
-import org.hibernate.annotations.ColumnDefault;
 import org.hibernate.annotations.Comment;
 
 import com.sparta.ticketauction.domain.place.entity.Place;
 import com.sparta.ticketauction.global.entity.BaseEntity;
 
-import jakarta.persistence.CascadeType;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.FetchType;
@@ -19,7 +15,6 @@ import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
 import jakarta.persistence.JoinColumn;
 import jakarta.persistence.ManyToOne;
-import jakarta.persistence.OneToMany;
 import jakarta.persistence.Table;
 import lombok.AccessLevel;
 import lombok.Builder;
@@ -27,21 +22,13 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 
 @Getter
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
 @Entity
-@Table(name = "goods")
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+@Table
 public class Goods extends BaseEntity {
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
 	private Long id;
-
-	@Comment("공연 제목")
-	@Column(name = "name", length = 30, nullable = false)
-	private String name;
-
-	@Comment("공연 내용")
-	@Column(name = "description", length = 150, nullable = false)
-	private String description;
 
 	@Comment("공연 시작일")
 	@Column(name = "start_date", nullable = false)
@@ -51,58 +38,22 @@ public class Goods extends BaseEntity {
 	@Column(name = "end_date", nullable = false)
 	private LocalDate endDate;
 
-	@Comment("공연 시간")
-	@Column(name = "running_time", nullable = false)
-	@ColumnDefault("0")
-	private Integer runningTime = 0;
-
-	@Comment("연령대")
-	@Column(name = "age_grade", nullable = false)
-	private AgeGrade ageGrade;
-
-	@Comment("공연 카테고리")
+	@Comment("공연")
 	@ManyToOne(fetch = FetchType.LAZY)
-	@JoinColumn(name = "goods_category_id")
-	private GoodsCategory goodsCategory;
+	@JoinColumn(name = "goods_info")
+	private GoodsInfo goodsInfo;
 
 	@Comment("공연장")
 	@ManyToOne(fetch = FetchType.LAZY)
 	@JoinColumn(name = "place_id")
 	private Place place;
 
-	@Comment("공연 이미지")
-	@OneToMany(mappedBy = "goods", cascade = CascadeType.ALL, orphanRemoval = true)
-	private List<GoodsImage> goodsImage = new ArrayList<>();
-
 	@Builder
-	private Goods(
-		String name,
-		String description,
-		LocalDate startDate,
-		LocalDate endDate,
-		Integer runningTime,
-		Integer ageGrade,
-		GoodsCategory goodsCategory,
-		List<GoodsImage> goodsImage,
-		Place place
-	) {
-		this.name = name;
-		this.description = description;
+	private Goods(LocalDate startDate, LocalDate endDate, GoodsInfo goodsInfo, Place place) {
 		this.startDate = startDate;
 		this.endDate = endDate;
-		this.runningTime = runningTime;
-		this.ageGrade = AgeGrade.of(ageGrade);
-		this.goodsCategory = goodsCategory;
-		this.goodsImage = goodsImage;
+		this.goodsInfo = goodsInfo;
 		this.place = place;
-	}
-
-	public void addGoodsImage(List<GoodsImage> goodsImages) {
-		this.goodsImage.addAll(goodsImages);
-	}
-
-	public void updateGoodsCategory(GoodsCategory goodsCategory) {
-		this.goodsCategory = goodsCategory;
 	}
 
 }
