+++ src/main/java/com/sparta/ticketauction/domain/goods/repository/GoodsRepository.java
@@ -2,7 +2,7 @@ package com.sparta.ticketauction.domain.goods.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import com.sparta.ticketauction.domain.goods.entity.Goods;
+import com.sparta.ticketauction.domain.goods.entity.GoodsInfo;
 
-public interface GoodsRepository extends JpaRepository<Goods, Long> {
+public interface GoodsRepository extends JpaRepository<GoodsInfo, Long> {
 }
