+++ src/main/java/com/sparta/ticketauction/domain/reservation/service/ReservationService.java
@@ -1,9 +1,5 @@
 package com.sparta.ticketauction.domain.reservation.service;
 
-import com.sparta.ticketauction.domain.reservation.request.ReservationCreateRequest;
-import com.sparta.ticketauction.domain.reservation.response.ReservationDetailResponse;
-import com.sparta.ticketauction.domain.user.entity.User;
-
 public interface ReservationService {
 
 	/**
@@ -14,10 +10,10 @@ public interface ReservationService {
 	 * @param request 예매 요청 정보
 	 * @return 예매한 좌석의 정보를 반환한다.
 	 */
-	ReservationDetailResponse reserve(
-		User user,
-		Long seatId,
-		Long sequenceId,
-		ReservationCreateRequest request
-	);
+	// ReservationDetailResponse reserve(
+	// 	User user,
+	// 	Long seatId,
+	// 	Long sequenceId,
+	// 	ReservationCreateRequest request
+	// );
 }
